Refactored product view to remote handler with subclasses
